Move is_awaiter<T> into cppcoro::detail namespace.
It is just an implementation detail of is_awaitable<T> and
shouldn't form part of the public API.
