+++ lib/build.cake
@@ -10,7 +10,6 @@ from cake.tools import compiler, script, env, project, variant
 includes = cake.path.join(env.expand('${CPPCORO}'), 'include', 'cppcoro', [
   'awaitable_traits.hpp',
   'is_awaitable.hpp',
-  'is_awaiter.hpp',
   'async_auto_reset_event.hpp',
   'async_manual_reset_event.hpp',
   'async_generator.hpp',
@@ -56,6 +55,7 @@ detailIncludes = cake.path.join(env.expand('${CPPCORO}'), 'include', 'cppcoro', 
   'when_all_counter.hpp',
   'when_all_task.hpp',
   'get_awaiter.hpp',
+  'is_awaiter.hpp',
   'any.hpp',
   'sync_wait_task.hpp',
   'unwrap_reference.hpp',
